Sets up with starter code/files
Wasted a ton of time in `lm01` with the client side doing things
they've done a million times, but typos and syntax errors cost us
dearly. To prevent that, I've added all the stuff they've done multiple
times already, and only have the crux of the lesson left to fill in.